third_party/WebKit/LayoutTests/svg/hixie/error/007-expected.txt

CONSOLE ERROR: line 4: Error: <rect> attribute width: A negative value is not valid. ("-200")
